requirements.txt
asgiref==3.5.2
coverage==6.4.2
Django==3.1.8
django-appconf==1.0.5
django-ckeditor==5.9.0
django-countries==7.3.2
django-country==0.1.0
django-imagekit==4.1.0
django-js-asset==2.0.0
pilkit==2.0
Pillow==9.2.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.3.0
tzdata==2022.1
